refactor: update import paths for types and dtos
