+++ src/infra/http/controllers/appointments/create-appointments.controller.ts
@@ -1,6 +1,6 @@
 import { Roles } from '@/core/shared/decorators/roles.decorator';
-import type { CreateAppointmentDto } from '@/core/shared/dto/appointments/create-appointment.dto';
-import { ROLE } from '@/core/shared/types';
+import type { CreateAppointmentDto } from '@/core/dto/appointments/create-appointment.dto';
+import { ROLE } from '@/core/types';
 import { CreateAppointmentsService } from '@/domain/services/appointments/create-appointments.service';
 import { Body, Controller, HttpCode, Post } from '@nestjs/common';
