+++ src/infra/http/controllers/user/delete-user-profile.controller.ts
@@ -13,8 +13,8 @@ import {
   ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
 import { Roles } from '@/core/shared/decorators/roles.decorator';
-import type { AuthenticatedUserRequest } from '@/core/shared/dto/user/get-user.dto';
-import { ROLE } from '@/core/shared/types';
+import type { AuthenticatedUserRequest } from '@/core/dto/user/get-user.dto';
+import { ROLE } from '@/core/types';
 import { DeleteUserProfileService } from '@Services/user/delete-user-profile.service';
 import { JwtAuthGuard } from '../../../auth/auth.guard';
