+++ src/infra/http/controllers/user/edit-user-profile.controller.ts
@@ -7,9 +7,9 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { Roles } from '@/core/shared/decorators/roles.decorator';
-import type { EditProfileDto } from '@/core/shared/dto/user/edit-profile.dto';
-import type { AuthenticatedUserRequest } from '@/core/shared/dto/user/get-user.dto';
-import { ROLE } from '@/core/shared/types';
+import type { EditProfileDto } from '@/core/dto/user/edit-profile.dto';
+import type { AuthenticatedUserRequest } from '@/core/dto/user/get-user.dto';
+import { ROLE } from '@/core/types';
 import { EditUserProfileService } from '@Services/user/edit-user-profile.service';
 import { JwtAuthGuard } from '@/infra/auth/auth.guard';
